REFACTOR: Changed _components to component
