+++ app/components/SponsorCard.tsx
