+++ app/components/home/FeaturedEvents.tsx
