+++ app/components/home/Hero.tsx
