+++ app/components/icons/AUECFooterLogo.tsx
