+++ app/layout.tsx
@@ -2,8 +2,8 @@ import type { Metadata } from "next";
 import Switzer from "next/font/local";
 import Syne from "next/font/local";
 import "./globals.css";
-import Navbar from "./_components/navbar/Navbar";
-import Footer from "./_components/footer/Footer";
+import Navbar from "./components/navbar/Navbar";
+import Footer from "./components/footer/Footer";
 import { ReactLenis } from "lenis/react";
 
 const switzer = Switzer({
